Renaming the folder from freqCount to validateInput
